feat: Allow to search images on google (#58)

+++ src/Services/storage.ts
@@ -2,6 +2,7 @@ import firebase from 'firebase';
 import * as cache from './cache';
 
 const storageRef = firebase.storage().ref();
+const isFirebaseFile = (urlOrPath: string) => !urlOrPath.startsWith('https');
 
 export const ITEMS_PATH = 'items';
 
@@ -17,14 +18,16 @@ export async function upload(fileName: string, file: ArrayBuffer) {
   return fullPath;
 }
 
-export function remove(filePath: string) {
+export async function remove(filePath: string) {
   cache.remove(getImageUrl(filePath));
-  const productImageRef = storageRef.child(filePath);
-  return productImageRef.delete();
+  if (isFirebaseFile(filePath)) {
+    const productImageRef = storageRef.child(filePath);
+    return productImageRef.delete();
+  }
 }
 
 export function getImageUrl(image: string) {
-  return `https://firebasestorage.googleapis.com/v0/b/sync-shopping-list-5e6ea.appspot.com/o/${encodeURIComponent(
+  return `https://firebasestorage.googleapis.com/v0/b/${process.env.REACT_APP_STORAGE_BUCKET}/o/${encodeURIComponent(
     image
   )}?alt=media`;
 }
